backend/requirements.txt
asgiref==3.9.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
cached-property==1.5.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
Django==5.2.6
django-allauth==65.16.1
django-axes==8.3.1
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
drf-spectacular==0.29.0
freetype-py==2.4.0
freezegun==1.2.1
gevent==24.2.1
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.4
idna==3.6
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libsass==0.22.0
louis==3.29.0
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
num2words==0.5.13
numpy==2.4.3
oauthlib==3.2.2
odoo==16.0
ofxparse==0.21
olefile==0.46
packaging==24.0
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
polib==1.1.1
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
pydot==1.4.2
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==2.12.1
pypng==0.20231004.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-decouple==3.8
python-dotenv==1.0.1
python-ldap==3.4.4
python-slugify==8.0.4
python-stdnum==1.19
pytz==2024.1
pyusb==1.2.1.post2
pyxdg==0.28
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.1.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.7.1
rlPyCairo==0.3.0
roman==3.3
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.3
systemd-python==235
text-unidecode==1.3
texttable==1.6.7
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
vobject==0.9.6.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
wheel==0.42.0
whitenoise==6.12.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
yt-dlp==2024.4.9
zeep==4.2.1
zope.event==5.0
zope.interface==6.1
